+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   API routes are treated the same as web routes.
 - Adjusted the template wrapper middleware to use block templates instead of
   `get_header()`/`get_footer()` calls.
+- Added callback support to the `assertContent()` method to allow for
+  more complex assertions on the content of a response.
 
 ### Fixed